requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
sendgrid==6.11.0
celery==5.4.0
redis==5.2.1